requirements.txt
aiohttp==3.9.1
beautifulsoup4==4.12.2
lxml==4.9.3
pytrends==4.9.2
pandas>=2.0.0
tweepy==4.14.0
python-dotenv==1.0.0
selenium==4.15.2
webdriver-manager==4.0.1
pyperclip==1.8.2

# Video creation dependencies
moviepy==1.0.3
requests==2.31.0
gtts==2.4.0
edge-tts==6.1.9
Pillow==10.1.0
imageio==2.33.0
imageio-ffmpeg==0.4.9

# Dashboard dependencies
flask==3.0.0
flask-cors==4.0.0

# YouTube upload dependencies
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Text-to-Speech dependencies
google-genai

# Bark TTS dependencies (Suno AI - high quality local TTS)
git+https://github.com/suno-ai/bark.git
scipy>=1.10.0
